Revert "debian/control: Add Build-Depends on libssh-dev for Ubuntu" (closes: #1002597)
This reverts commit 59825944dede2c1a3bf9ddab7c84b62e8d3586e4.

The change had side effects on Debian due to the inclusion of the new
Build-dep, even though it doesn't changes the resulting binary. It cause
issues for architecture bootstraping.

We are gonna reintroduce this change once the issues are fixed, to allow
Ubuntu to remove its delta.

See discussions at #1002598 and #1002597 for details